requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
install==1.3.5
Markdown==3.3.6
parts==1.3.0
postgres==4.0
products==0.3.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2021.5
zipp==3.8.0
